+++ go/internal/types/extension_types.go
@@ -22,7 +22,6 @@ import (
 	"encoding/binary"
 	"fmt"
 	"reflect"
-	"strconv"
 	"strings"
 
 	"github.com/apache/arrow/go/v18/arrow"
@@ -31,8 +30,6 @@ import (
 	"github.com/google/uuid"
 	"golang.org/x/xerrors"
 )
-
-const ExtensionNameBool8 = "bool8"
 
 var UUID = NewUUIDType()
 
@@ -537,204 +534,15 @@ func (SmallintType) Deserialize(storageType arrow.DataType, data string) (arrow.
 	return NewSmallintType(), nil
 }
 
-type Bool8Type struct {
-	arrow.ExtensionBase
-}
-
-// ArrayType implements arrow.ExtensionType.
-func (b *Bool8Type) ArrayType() reflect.Type {
-	return reflect.TypeOf(Bool8Array{})
-}
-
-// Deserialize implements arrow.ExtensionType.
-func (b *Bool8Type) Deserialize(storageType arrow.DataType, data string) (arrow.ExtensionType, error) {
-	if data != ExtensionNameBool8 {
-		return nil, fmt.Errorf("type identifier did not match: '%s'", data)
-	}
-	if !arrow.TypeEqual(storageType, arrow.PrimitiveTypes.Uint8) {
-		return nil, fmt.Errorf("invalid storage type for Bool8Type: %s", storageType.Name())
-	}
-	return NewBool8Type(), nil
-}
-
-// ExtensionEquals implements arrow.ExtensionType.
-func (b *Bool8Type) ExtensionEquals(other arrow.ExtensionType) bool {
-	return b.ExtensionName() == other.ExtensionName()
-}
-
-// ExtensionName implements arrow.ExtensionType.
-func (b *Bool8Type) ExtensionName() string {
-	return ExtensionNameBool8
-}
-
-// Serialize implements arrow.ExtensionType.
-func (b *Bool8Type) Serialize() string {
-	return ExtensionNameBool8
-}
-
-// String implements arrow.ExtensionType.
-// Subtle: this method shadows the method (ExtensionBase).String of Bool8Type.ExtensionBase.
-func (b *Bool8Type) String() string {
-	panic("unimplemented")
-}
-
-func NewBool8Type() *Bool8Type {
-	return &Bool8Type{ExtensionBase: arrow.ExtensionBase{
-		Storage: arrow.PrimitiveTypes.Uint8}}
-}
-
-func (*Bool8Type) NewBuilder(bldr *array.ExtensionBuilder) array.Builder {
-	return NewBool8Builder(bldr)
-}
-
-type Bool8Array struct {
-	array.ExtensionArrayBase
-}
-
-func (a *Bool8Array) String() string {
-	var o strings.Builder
-	arr := a.Storage().(*array.Uint8)
-	o.WriteString("[")
-	for i := 0; i < arr.Len(); i++ {
-		if i > 0 {
-			o.WriteString(" ")
-		}
-		switch {
-		case a.IsNull(i):
-			o.WriteString(array.NullValueStr)
-		default:
-			fmt.Fprintf(&o, "%v", a.Value(i))
-		}
-	}
-	o.WriteString("]")
-	return o.String()
-}
-
-func (a *Bool8Array) Value(i int) bool {
-	return uint8ToBool(a.Storage().(*array.Uint8).Value(i))
-}
-
-func (a *Bool8Array) ValueStr(i int) string {
-	switch {
-	case a.IsNull(i):
-		return array.NullValueStr
-	default:
-		return fmt.Sprint(a.Value(i))
-	}
-}
-
-func (a *Bool8Array) MarshalJSON() ([]byte, error) {
-	values := make([]interface{}, a.Len())
-	for i := 0; i < a.Len()-a.NullN(); i++ {
-		if a.IsValid(i) {
-			values[i] = a.Value(i)
-		}
-	}
-	return json.Marshal(values)
-}
-
-func (a *Bool8Array) GetOneForMarshal(i int) interface{} {
-	if a.IsNull(i) {
-		return nil
-	}
-	return a.Value(i)
-}
-
-func boolToUint8(v bool) uint8 {
-	var res uint8
-	if v {
-		res = 1
-	}
-	return res
-}
-
-func uint8ToBool(v uint8) bool {
-	return v != 0
-}
-
-type Bool8Builder struct {
-	*array.ExtensionBuilder
-}
-
-func NewBool8Builder(builder *array.ExtensionBuilder) *Bool8Builder {
-	builder.Retain()
-	return &Bool8Builder{ExtensionBuilder: builder}
-}
-
-func (b *Bool8Builder) Append(v bool) {
-	b.ExtensionBuilder.Builder.(*array.Uint8Builder).Append(boolToUint8(v))
-}
-
-func (b *Bool8Builder) UnsafeAppend(v bool) {
-	b.ExtensionBuilder.Builder.(*array.Uint8Builder).UnsafeAppend(boolToUint8(v))
-}
-
-func (b *Bool8Builder) AppendValueFromString(s string) error {
-	if s == array.NullValueStr {
-		b.AppendNull()
-		return nil
-	}
-
-	val, err := strconv.ParseBool(s)
-	if err != nil {
-		return err
-	}
-
-	b.Append(val)
-	return nil
-}
-
-func (b *Bool8Builder) AppendValues(v []bool, valid []bool) {
-	boolsAsBytes := arrow.GetBool8Bytes(v)
-	boolsAsUint8s := arrow.GetData[uint8](boolsAsBytes)
-	b.ExtensionBuilder.Builder.(*array.Uint8Builder).AppendValues(boolsAsUint8s, valid)
-}
-
-func (b *Bool8Builder) UnmarshalOne(dec *json.Decoder) error {
-	t, err := dec.Token()
-	if err != nil {
-		return err
-	}
-
-	switch v := t.(type) {
-	case bool:
-		b.Append(v)
-		return nil
-	case string:
-		return b.AppendValueFromString(v)
-	case nil:
-		b.AppendNull()
-		return nil
-	default:
-		return &json.UnmarshalTypeError{
-			Value:  fmt.Sprint(t),
-			Type:   reflect.TypeOf([]byte{}),
-			Offset: dec.InputOffset(),
-			Struct: fmt.Sprintf("FixedSizeBinary[%d]", 16),
-		}
-	}
-}
-
-func (b *Bool8Builder) Unmarshal(dec *json.Decoder) error {
-	for dec.More() {
-		if err := b.UnmarshalOne(dec); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
 var (
 	_ arrow.ExtensionType  = (*Parametric1Type)(nil)
 	_ arrow.ExtensionType  = (*Parametric2Type)(nil)
 	_ arrow.ExtensionType  = (*ExtStructType)(nil)
 	_ arrow.ExtensionType  = (*DictExtensionType)(nil)
 	_ arrow.ExtensionType  = (*SmallintType)(nil)
-	_ arrow.ExtensionType  = (*Bool8Type)(nil)
 	_ array.ExtensionArray = (*Parametric1Array)(nil)
 	_ array.ExtensionArray = (*Parametric2Array)(nil)
 	_ array.ExtensionArray = (*ExtStructArray)(nil)
 	_ array.ExtensionArray = (*DictExtensionArray)(nil)
 	_ array.ExtensionArray = (*SmallintArray)(nil)
-	_ array.ExtensionArray = (*Bool8Array)(nil)
 )
